+++ CHANGELOG.md
@@ -23,10 +23,14 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Fetch documents from backend and download
    * Update style multiple order section
    * Add short description to card
+* AppReleaseProcess
+   * Conformity Document Type - Add Dropzone & Document handling
 * App Admin Board
    * Add Section for Conformity Document & display via api
 * AppOverview
    * Page Load handling
+* Organization
+   * Active subscription image issue fix
 
 ## 1.0.0-RC10